+++ website/src/components/User/AvatarBadge.tsx
@@ -2,17 +2,16 @@ import AccountCircleIcon from '~icons/ic/baseline-account-circle';
 import type { FC } from 'react';
 
 type Props = {
-    username?: string;
+    displayName?: string;
 };
 
-export const AvatarBadge: FC<Props> = (props) => {
-    const { username } = props || {};
+export const AvatarBadge: FC<Props> = ({ displayName }) => {
 
     return (
         <div className='flex flex-col items-center justify-center'>
             <AccountCircleIcon fontSize={100} />
             <div className='flex flex-col items-left justify-center'>
-                <div className='text-lg'>{username ?? ''}</div>
+                <div className='text-lg'>{displayName ?? ''}</div>
             </div>
         </div>
     );
