+++ website/src/pages/datasets/[datasetId].astro
@@ -13,8 +13,8 @@ const clientConfig = getRuntimeConfig().public;
 const accessToken = getAccessToken(Astro.locals.session)!;
 const { datasetId } = Astro.params;
 const version = Astro.url.searchParams.get('version')! || '';
-const name = Astro.locals.session.user!.name;
-const username = Astro.locals.session.user!.username;
+
+const username = Astro.locals.session.user?.username;
 
 const datasetResponse = await BackendClient.create().astroFileTypeHelpers.getDataset(
     accessToken,
@@ -30,14 +30,14 @@ const datasetRecordsResponse = await BackendClient.create().astroFileTypeHelpers
 ---
 
 <BaseLayout title='Datasets' data-testid='datasets-item-container'>
-    <BackButton marginRight={0} client:only='react' />
+    <BackButton marginRight={0} client:load />
     <div class='flex flex-col justify-start'>
         {
             datasetResponse.match(
                 (dataset) => (
                     <div class='flex flex-row items-left'>
                         <div class='w-1/4 flex flex-col justify-start items-center'>
-                            <AvatarBadge username={name} />
+                            <AvatarBadge displayName={dataset.createdBy} />
                         </div>
                         <div class='w-3/4'>
                             {
